Fix: Type issue in Auth-form & logic

+++ src/actions/action.ts
@@ -2,17 +2,26 @@
 
 import { signIn, signOut } from "@/lib/auth";
 import prisma from "@/lib/db";
-import { petFormSchema, petIdSchema } from "@/lib/schema";
+import { authSchema, petFormSchema, petIdSchema } from "@/lib/schema";
 import { handleErrors, sleep } from "@/lib/utils";
 import { revalidatePath } from "next/cache";
 import bcrypt from "bcryptjs";
+import { checkAuth, getPetById } from "@/lib/server-utils";
+import { redirect } from "next/navigation";
 
-// ------ AUTH actions --------
+//  ------------------------------------------------------
+// 						AUTH actions
+// --------------------------------------------------------
 
-export async function login(formData: FormData) {
-	const authData = Object.fromEntries(formData.entries());
+export async function login(formData: unknown) {
+	// Check if form data is valid
+	if (!(formData instanceof FormData)) {
+		return { message: "Invalid form data." };
+	}
 
-	await signIn("credentials", authData);
+	await signIn("credentials", formData);
+
+	redirect("/app/dashboard");
 }
 
 export async function logout() {
@@ -20,10 +29,21 @@ export async function logout() {
 }
 
 export async function signUp(formData: FormData) {
-	const authData = Object.fromEntries(formData.entries());
-	const { email, password } = authData;
+	// Check if form data is valid
+	if (!(formData instanceof FormData)) {
+		return { message: "Invalid form data." };
+	}
 
-	const salt = bcrypt.genSaltSync(10);
+	const authData = Object.fromEntries(formData.entries());
+	const validatedData = authSchema.safeParse(authData);
+
+	if (!validatedData.success) {
+		return { message: "Invalid form data." };
+	}
+
+	const { email, password } = validatedData.data;
+
+	const salt = await bcrypt.genSalt(10);
 	const hashedPassword = await bcrypt.hash(password, salt);
 
 	try {
@@ -41,10 +61,13 @@ export async function signUp(formData: FormData) {
 	await signIn("credentials", authData);
 }
 
-// ------- PET actions ---------
+//  ------------------------------------------------------
+// 						PET actions
+// --------------------------------------------------------
 
 export async function addPet(pet: unknown) {
-	await sleep(2000);
+	await sleep(1000);
+	const session = await checkAuth();
 
 	const validatedPet = petFormSchema.safeParse(pet);
 
@@ -56,7 +79,14 @@ export async function addPet(pet: unknown) {
 
 	try {
 		await prisma.pet.create({
-			data: validatedPet.data,
+			data: {
+				...validatedPet.data,
+				owner: {
+					connect: {
+						id: session.user.id,
+					},
+				},
+			},
 		});
 	} catch (error: unknown) {
 		const message = handleErrors(error, "Could not add pet.");
@@ -67,7 +97,8 @@ export async function addPet(pet: unknown) {
 }
 
 export async function editPet(petId: unknown, pet: unknown) {
-	await sleep(2000);
+	await sleep(1000);
+	const session = await checkAuth();
 
 	const validatedPetID = petIdSchema.safeParse(petId);
 	const validatedPet = petFormSchema.safeParse(pet);
@@ -78,6 +109,22 @@ export async function editPet(petId: unknown, pet: unknown) {
 		};
 	}
 
+	// authorization check
+	const userPet = await getPetById(validatedPetID.data);
+
+	if (!userPet) {
+		return {
+			message: "Could not find pet.",
+		};
+	}
+
+	if (userPet?.ownerId !== session.user.id) {
+		return {
+			message: "You are not authorized to edi this pet!",
+		};
+	}
+
+	// Update data from db
 	try {
 		await prisma.pet.update({
 			where: {
@@ -94,6 +141,9 @@ export async function editPet(petId: unknown, pet: unknown) {
 }
 
 export async function deletePet(petId: unknown) {
+	await sleep(1000);
+	const session = await checkAuth();
+
 	const validatedPetID = petIdSchema.safeParse(petId);
 
 	if (!validatedPetID.success) {
@@ -102,6 +152,22 @@ export async function deletePet(petId: unknown) {
 		};
 	}
 
+	// authorization check
+	const pet = await getPetById(validatedPetID.data);
+
+	if (!pet) {
+		return {
+			message: "Could not find pet.",
+		};
+	}
+
+	if (pet?.ownerId !== session.user.id) {
+		return {
+			message: "You are not authorized to delete this pet!",
+		};
+	}
+
+	// Delete data from DB
 	try {
 		await prisma.pet.delete({
 			where: {
